requirements_test.txt
-r requirements_dev.txt
pytest-homeassistant-custom-component==0.1.0
proconip==1.1.0
